requirements.txt
    pydub
    Flask
    requests    
    requests_to_curl
    langchain
    ffmpeg-python
    wikipedia
    openai 
    google-api-python-client
    oauth2client==3.0.0